fix(ci): extract CLI skip pytest plugin and format
Keep tests/conftest.py under LOC guard; register tests.plugin_cli_skip via pyproject addopts.
